Remove default policy on agent network onboarding
